# api_gateway/api/docs/swagger.yaml
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  schedule_service.CreateLesson:
    properties:
      lesson:
        type: string
      schedule_id:
        type: string
    type: object
...
  schedule_service.GetListLessonResponse:
    properties:
      Lessons:
        items:
          $ref: '#/definitions/schedule_service.Lesson'
        type: array
      count:
        type: integer
    type: object
...
  schedule_service.Lesson:
    properties:
      Tasks:
        items:
          $ref: '#/definitions/schedule_service.Task'
        type: array
      created_at:
        type: string
      id:
        type: string
      lesson:
        type: string
      schedule_id:
        type: string
      updated_at:
        type: string
    type: object
...
  schedule_service.Task:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: string
      label:
        type: string
      lesson_id:
        type: string
      score:
        type: integer
      updated_at:
        type: string
    type: object
...
  schedule_service.UpdateLesson:
    properties:
      id:
        type: string
      lesson:
        type: string
      schedule_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /lesson:
    get:
      consumes:
      - application/json
      description: Get Lessons List
      operationId: get_lesson_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllLessonResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.GetListLessonResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Lessons List
      tags:
      - Lesson
    post:
      consumes:
      - application/json
      description: Create Lesson
      operationId: create_lesson
      parameters:
      - description: CreateLessonBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.CreateLesson'
      produces:
      - application/json
      responses:
        "200":
          description: GetLessonBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.Lesson'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Lesson
      tags:
      - Lesson
  /lesson/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Lesson
      operationId: delete_lesson
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lesson data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Lesson
      tags:
      - Lesson
    get:
      consumes:
      - application/json
      description: Get Lesson  By ID
      operationId: get_lesson_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: LessonBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.Lesson'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Lesson  By ID
      tags:
      - Lesson
    put:
      consumes:
      - application/json
      description: Update Lesson
      operationId: update_lesson
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateLesson
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schedule_service.UpdateLesson'
      produces:
      - application/json
      responses:
        "200":
          description: Lesson data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/schedule_service.Lesson'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Lesson
      tags:
      - Lesson
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"